[HWKINVENT-57] Get rid of the inventory mixins. That was just a bad idea.
Also, there is just a single "@Official" inventory instance to be
shared in CDI. REST builds on top of that to build a custom
"AutoTenant" implementation that suites the needs of the security
imposed by the REST API. AutoTenant is not needed anywhere else but
the REST api so it made little sense to have it right in the
component API.
