--- hawkular-inventory-cdi/src/main/java/org/hawkular/inventory/cdi/AutoTenantInventoryInitialized.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.inventory.cdi;
-
-import org.hawkular.inventory.api.Inventory;
-
-/**
- * A CDI event informing the users that an auto-tenant inventory instance was initialized.
- *
- * @author Lukas Krejci
- * @since 0.0.2
- */
-public class AutoTenantInventoryInitialized extends AbstractInventoryInitializedEvent<Inventory.Mixin.AutoTenant> {
-    public AutoTenantInventoryInitialized(Inventory.Mixin.AutoTenant inventory) {
-        super(inventory);
-    }
-}
