+++ rest-servlet/src/main/java/org/hawkular/inventory/rest/RestBase.java
@@ -18,7 +18,6 @@ package org.hawkular.inventory.rest;
 
 import org.hawkular.accounts.api.PersonaService;
 import org.hawkular.inventory.api.Inventory;
-import org.hawkular.inventory.cdi.AutoTenant;
 
 import javax.inject.Inject;
